refactor: 강의 및 과제 관련 타입 수정 및 API 호출 인자 변경 #55
