+++ types/lecture.ts
@@ -1,0 +1,10 @@
+export interface Lecture {
+  id: number;
+  name: string;
+  description: string;
+  url: string;
+  created_at: string;
+  challenge_lecture_id: number;
+  open_at: string;
+  challenge_id: string;
+}
